Split devices and backends into their own sub-spec types
