+++ lib/propolis-client/src/instance_spec/mod.rs
@@ -6,16 +6,29 @@
 //!
 //! # Spec format
 //!
-//! Instance specs describe a VM's virtual mainboard (its CPUs, memory, chipset,
-//! and platform details) and collections of virtual hardware components. Some
-//! components, such as serial ports, are freestanding; others are split into a
-//! "device" frontend, which specifies the virtual hardware interface exposed to
-//! guest software, and a backend that provides services (e.g. durable storage
-//! or a network interface) to devices from the host system and/or the rest of
-//! the rack.
+//! Instance specs are divided into two parts:
 //!
-//! Devices and backends are named. Collections of components of a given type
-//! are stored in maps from names to component descriptions.
+//! - The "device" half describes the VM components and interfaces that guest
+//!   software can observe directly. Device configuration generally can't change
+//!   at runtime without the guest's cooperation.
+//! - The "backend" half describes how the VM connects to services (provided by
+//!   the host OS or other parts of the rack) that supply functions the devices
+//!   need to provide their abstractions to guests.
+//!
+//! For example, to expose a virtual NVMe disk to a guest, a spec defines an
+//! NVMe device (expressing that the VMM should create a PCI device exposing an
+//! NVMe-conforming interface at a specific bus/device/function) and connects it
+//! to a storage backend (expressing that I/O to the virtual disk should be
+//! serviced by a local file, or by the Crucible storage service, or by a buffer
+//! in the VMM's memory).
+//!
+//! # Instance specs and the VM lifecycle
+//!
+//! Instance specs are used to initialize new VMs and during live migration.
+//! VM initialization uses specs to determine what components to create. Live
+//! migration uses the [`MigrationCompatible`] trait to determine whether two
+//! specs would, if realized, create VMs that are sufficiently compatible to
+//! allow one VM to migrate to the other.
 //!
 //! # Verification
 //!
@@ -27,8 +40,7 @@
 //!
 //! # Versioning
 //!
-//! Instance spec versioning is not fully formalized yet. RFD 283 discusses some
-//! possible approaches to formalizing it.
+//! NOTE: Instance spec versioning is not fully formalized yet; see RFD 283.
 //!
 //! What versioning requirements exist today are enforced through serde
 //! attributes:
@@ -45,13 +57,18 @@
 //! field.
 
 use std::collections::BTreeMap;
-use std::convert::TryFrom;
 
 use serde::{Deserialize, Serialize};
 use thiserror::Error;
 
 pub use crucible_client_types::VolumeConstructionRequest;
 pub use propolis_types::PciPath;
+
+mod backends;
+mod devices;
+
+pub use backends::*;
+pub use devices::*;
 
 /// Type alias for keys in the instance spec's maps.
 type SpecKey = String;
@@ -184,409 +201,6 @@ impl<T: MigrationCompatible> MigrationCompatible for BTreeMap<SpecKey, T> {
     }
 }
 
-/// A kind of virtual chipset.
-#[derive(Clone, Copy, Deserialize, Serialize, Debug, PartialEq, Eq)]
-#[serde(deny_unknown_fields)]
-pub enum Chipset {
-    /// An Intel 440FX-compatible chipset.
-    I440Fx {
-        /// Specifies whether the chipset should allow PCI configuration space
-        /// to be accessed through the PCIe extended configuration mechanism.
-        enable_pcie: bool,
-    },
-}
-
-/// A VM's mainboard.
-#[derive(Clone, Deserialize, Serialize, Debug, PartialEq, Eq)]
-#[serde(deny_unknown_fields)]
-pub struct Board {
-    /// The number of virtual logical processors attached to this VM.
-    pub cpus: u8,
-
-    /// The amount of guest RAM attached to this VM.
-    pub memory_mb: u64,
-
-    /// The chipset to expose to guest software.
-    pub chipset: Chipset,
-    // TODO: Guest platform and CPU feature identification.
-    // TODO: NUMA topology.
-}
-
-impl Default for Board {
-    fn default() -> Self {
-        Self {
-            cpus: 0,
-            memory_mb: 0,
-            chipset: Chipset::I440Fx { enable_pcie: false },
-        }
-    }
-}
-
-impl MigrationCompatible for Board {
-    fn is_migration_compatible(
-        &self,
-        other: &Self,
-    ) -> Result<(), SpecMismatchDetails> {
-        if self.cpus != other.cpus {
-            return Err(SpecMismatchDetails::CpuCount(self.cpus, other.cpus));
-        }
-        if self.memory_mb != other.memory_mb {
-            return Err(SpecMismatchDetails::MemorySize(
-                self.memory_mb,
-                other.memory_mb,
-            ));
-        }
-        match (self.chipset, other.chipset) {
-            (
-                Chipset::I440Fx { enable_pcie: this_enable },
-                Chipset::I440Fx { enable_pcie: other_enable },
-            ) => {
-                if this_enable != other_enable {
-                    return Err(SpecMismatchDetails::PcieEnablement(
-                        this_enable,
-                        other_enable,
-                    ));
-                }
-            }
-        }
-
-        Ok(())
-    }
-}
-
-//
-// Storage devices.
-//
-
-/// A description of a Crucible volume construction request.
-#[derive(Clone, Deserialize, Serialize, Debug)]
-pub struct CrucibleRequestContents {
-    /// A [`crucible_client_types::VolumeConstructionRequest`], serialized as JSON.
-    //
-    // Storing volume construction requests in serialized form allows external
-    // types to change without causing a breaking change to instance specs.
-    // Consider the following scenario, assuming the VolumeConstructionRequest
-    // struct is used directly:
-    //
-    // - Sled agent v1 starts Propolis v1 using Crucible request v1.
-    // - Sled agent v2 (on some other sled with newer software) starts Propolis
-    //   v2 using Crucible request v2, which has a new field not present in v1.
-    // - Nexus orders a migration from v1 to v2. This requires someone to
-    //   compare the two instances' specs for migratability.
-    //
-    // Migration compatibility is normally checked by the two Propolis servers
-    // involved: one server sends its instance spec to the other, and the
-    // recipient compares the specs to see if they're compatible. In this case,
-    // v2 can't deserialize v1's spec (a field is missing), and v1 can't
-    // deserialize v2's (an extra field is present), so migration will always
-    // fail.
-    //
-    // Storing a serialized request avoids this problem as follows:
-    //
-    // - Sled agent v2 starts Propolis v2 with spec v2. It deserializes the
-    //   request contents in the spec body into a v2 construction request.
-    // - Migration begins. Propolis v2 can now deserialize the v1 instance spec
-    //   and check it for compatibility. It can't deserialize the v1 *request
-    //   contents*, but this can be dealt with separately (e.g. by having the v1
-    //   and v2 Crucible components in the Propolis server negotiate
-    //   compatibility themselves, which is an affordance the migration protocol
-    //   allows).
-    pub json: String,
-}
-
-impl TryFrom<&CrucibleRequestContents>
-    for crucible_client_types::VolumeConstructionRequest
-{
-    type Error = serde_json::Error;
-
-    fn try_from(value: &CrucibleRequestContents) -> Result<Self, Self::Error> {
-        serde_json::from_str(&value.json)
-    }
-}
-
-impl TryFrom<&crucible_client_types::VolumeConstructionRequest>
-    for CrucibleRequestContents
-{
-    type Error = serde_json::Error;
-
-    fn try_from(
-        value: &crucible_client_types::VolumeConstructionRequest,
-    ) -> Result<Self, Self::Error> {
-        Ok(Self { json: serde_json::to_string(value)? })
-    }
-}
-
-/// A kind of storage backend: a connection to on-sled resources or other
-/// services that provide the functions storage devices need to implement their
-/// contracts.
-#[derive(Clone, Deserialize, Serialize, Debug)]
-#[serde(deny_unknown_fields)]
-pub enum StorageBackendKind {
-    /// A Crucible-backed device, containing a generation number and a
-    /// construction request.
-    Crucible { gen: u64, req: CrucibleRequestContents },
-
-    /// A device backed by a file on the host machine. The payload is a path to
-    /// this file.
-    File { path: String },
-
-    /// A device backed by an in-memory buffer in the VMM process. The initial
-    /// contents of this buffer are supplied out-of-band, either at
-    /// instance initialization time or from a migration source.
-    InMemory,
-}
-
-impl MigrationCompatible for StorageBackendKind {
-    fn is_migration_compatible(
-        &self,
-        other: &Self,
-    ) -> Result<(), SpecMismatchDetails> {
-        // Two storage backends are compatible if they have the same kind,
-        // irrespective of their configurations. The migration protocol allows
-        // individual backend instances to include messages in the preamble that
-        // allow a source and target to decide independently whether they are
-        // compatible with each other.
-        if std::mem::discriminant(self) != std::mem::discriminant(other) {
-            return Err(SpecMismatchDetails::StorageBackendKind(
-                self.clone(),
-                other.clone(),
-            ));
-        }
-
-        Ok(())
-    }
-}
-
-/// A storage backend.
-#[derive(Clone, Deserialize, Serialize, Debug)]
-#[serde(deny_unknown_fields)]
-pub struct StorageBackend {
-    /// The kind of storage backend this is.
-    pub kind: StorageBackendKind,
-
-    /// Whether the storage is read-only.
-    pub readonly: bool,
-}
-
-impl MigrationCompatible for StorageBackend {
-    fn is_migration_compatible(
-        &self,
-        other: &Self,
-    ) -> Result<(), SpecMismatchDetails> {
-        if self.readonly != other.readonly {
-            return Err(SpecMismatchDetails::StorageBackendReadonly(
-                self.readonly,
-                other.readonly,
-            ));
-        }
-        self.kind.is_migration_compatible(&other.kind)
-    }
-}
-
-/// A kind of storage device: the sort of virtual device interface the VMM
-/// exposes to guest software.
-#[derive(Clone, Copy, Deserialize, Serialize, Debug, PartialEq, Eq)]
-#[serde(deny_unknown_fields)]
-pub enum StorageDeviceKind {
-    Virtio,
-    Nvme,
-}
-
-/// A storage device.
-#[derive(Clone, Deserialize, Serialize, Debug, PartialEq, Eq)]
-#[serde(deny_unknown_fields)]
-pub struct StorageDevice {
-    /// The device interface to present to the guest.
-    pub kind: StorageDeviceKind,
-
-    /// The name of the device's backend.
-    pub backend_name: String,
-
-    /// The PCI path at which to attach this device.
-    pub pci_path: PciPath,
-}
-
-impl MigrationCompatible for StorageDevice {
-    fn is_migration_compatible(
-        &self,
-        other: &Self,
-    ) -> Result<(), SpecMismatchDetails> {
-        if self.kind != other.kind {
-            return Err(SpecMismatchDetails::StorageDeviceKind(
-                self.kind, other.kind,
-            ));
-        }
-        if self.backend_name != other.backend_name {
-            return Err(SpecMismatchDetails::StorageDeviceBackend(
-                self.backend_name.clone(),
-                other.backend_name.clone(),
-            ));
-        }
-        if self.pci_path != other.pci_path {
-            return Err(SpecMismatchDetails::PciPath(
-                self.pci_path,
-                other.pci_path,
-            ));
-        }
-        Ok(())
-    }
-}
-
-//
-// Network devices.
-//
-
-/// A kind of network backend: a connection to an on-sled networking resource
-/// that provides the functions needed for guest network adapters to implement
-/// their contracts.
-#[derive(Clone, Deserialize, Serialize, Debug)]
-#[serde(deny_unknown_fields)]
-pub enum NetworkBackendKind {
-    /// A virtio-net (viona) backend associated with the supplied named vNIC on
-    /// the host.
-    Virtio { vnic_name: String },
-}
-
-impl MigrationCompatible for NetworkBackendKind {
-    fn is_migration_compatible(
-        &self,
-        other: &Self,
-    ) -> Result<(), SpecMismatchDetails> {
-        // Two network backends are compatible if they have the same kind,
-        // irrespective of their configurations.
-        if std::mem::discriminant(self) != std::mem::discriminant(other) {
-            return Err(SpecMismatchDetails::NetworkBackendKind(
-                self.clone(),
-                other.clone(),
-            ));
-        }
-
-        Ok(())
-    }
-}
-
-/// A network backend.
-#[derive(Clone, Deserialize, Serialize, Debug)]
-#[serde(deny_unknown_fields)]
-pub struct NetworkBackend {
-    pub kind: NetworkBackendKind,
-}
-
-impl MigrationCompatible for NetworkBackend {
-    fn is_migration_compatible(
-        &self,
-        other: &Self,
-    ) -> Result<(), SpecMismatchDetails> {
-        self.kind.is_migration_compatible(&other.kind)
-    }
-}
-
-/// A virtual network adapter that presents a virtio network device interface.
-#[derive(Clone, Deserialize, Serialize, Debug, PartialEq, Eq)]
-#[serde(deny_unknown_fields)]
-pub struct NetworkDevice {
-    /// The name of the device's backend.
-    pub backend_name: String,
-
-    /// The PCI path at which to attach this device.
-    pub pci_path: PciPath,
-}
-
-impl MigrationCompatible for NetworkDevice {
-    fn is_migration_compatible(
-        &self,
-        other: &Self,
-    ) -> Result<(), SpecMismatchDetails> {
-        if self.backend_name != other.backend_name {
-            return Err(SpecMismatchDetails::NetworkDeviceBackend(
-                self.backend_name.clone(),
-                other.backend_name.clone(),
-            ));
-        }
-        if self.pci_path != other.pci_path {
-            return Err(SpecMismatchDetails::PciPath(
-                self.pci_path,
-                other.pci_path,
-            ));
-        }
-
-        Ok(())
-    }
-}
-
-//
-// Serial ports.
-//
-
-/// A serial port identifier, which determines what I/O ports a guest can use to
-/// access a port.
-#[derive(Clone, Copy, Deserialize, Serialize, Debug, PartialEq, Eq)]
-#[serde(deny_unknown_fields)]
-pub enum SerialPortNumber {
-    Com1,
-    Com2,
-    Com3,
-    Com4,
-}
-
-/// A serial port device.
-#[derive(Clone, Copy, Deserialize, Serialize, Debug, PartialEq, Eq)]
-#[serde(deny_unknown_fields)]
-pub struct SerialPort {
-    /// The serial port number for this port.
-    pub num: SerialPortNumber,
-}
-
-impl MigrationCompatible for SerialPort {
-    fn is_migration_compatible(
-        &self,
-        other: &Self,
-    ) -> Result<(), SpecMismatchDetails> {
-        if self.num != other.num {
-            return Err(SpecMismatchDetails::SerialPortNumber(
-                self.num, other.num,
-            ));
-        }
-
-        Ok(())
-    }
-}
-
-/// A PCI-PCI bridge.
-#[derive(Clone, Copy, Deserialize, Serialize, Debug, PartialEq, Eq)]
-#[serde(deny_unknown_fields)]
-pub struct PciPciBridge {
-    /// The logical bus number of this bridge's downstream bus. Other devices
-    /// may use this bus number in their PCI paths to indicate they should be
-    /// attached to this bridge's bus.
-    pub downstream_bus: u8,
-
-    /// The PCI path at which to attach this bridge.
-    pub pci_path: PciPath,
-}
-
-impl MigrationCompatible for PciPciBridge {
-    fn is_migration_compatible(
-        &self,
-        other: &Self,
-    ) -> Result<(), SpecMismatchDetails> {
-        if self.downstream_bus != other.downstream_bus {
-            return Err(SpecMismatchDetails::PciBridgeDownstreamBus(
-                self.downstream_bus,
-                other.downstream_bus,
-            ));
-        }
-        if self.pci_path != other.pci_path {
-            return Err(SpecMismatchDetails::PciPath(
-                self.pci_path,
-                other.pci_path,
-            ));
-        }
-
-        Ok(())
-    }
-}
-
 /// A full instance specification. See the documentation for individual
 /// elements for more information about the fields in this structure.
 ///
@@ -595,14 +209,8 @@ impl MigrationCompatible for PciPciBridge {
 #[derive(Default, Clone, Deserialize, Serialize, Debug)]
 #[serde(deny_unknown_fields)]
 pub struct InstanceSpec {
-    pub board: Board,
-
-    pub storage_devices: BTreeMap<SpecKey, StorageDevice>,
-    pub storage_backends: BTreeMap<SpecKey, StorageBackend>,
-    pub network_devices: BTreeMap<SpecKey, NetworkDevice>,
-    pub network_backends: BTreeMap<SpecKey, NetworkBackend>,
-    pub serial_ports: BTreeMap<SpecKey, SerialPort>,
-    pub pci_pci_bridges: BTreeMap<SpecKey, PciPciBridge>,
+    pub devices: DeviceSpec,
+    pub backends: BackendSpec,
 }
 
 impl InstanceSpec {
@@ -610,63 +218,13 @@ impl InstanceSpec {
         &self,
         other: &Self,
     ) -> Result<(), MigrationCompatibilityError> {
-        self.board.is_migration_compatible(&other.board).map_err(|e| {
-            MigrationCompatibilityError::SpecMismatch("board".to_string(), e)
-        })?;
+        self.devices
+            .is_migration_compatible(&other.devices)
+            .map_err(|e| MigrationCompatibilityError::SpecMismatch(e.0, e.1))?;
 
-        self.storage_devices
-            .is_migration_compatible(&other.storage_devices)
-            .map_err(|e| {
-                MigrationCompatibilityError::SpecMismatch(
-                    "storage devices".to_string(),
-                    e,
-                )
-            })?;
-
-        self.storage_backends
-            .is_migration_compatible(&other.storage_backends)
-            .map_err(|e| {
-                MigrationCompatibilityError::SpecMismatch(
-                    "storage backends".to_string(),
-                    e,
-                )
-            })?;
-
-        self.network_devices
-            .is_migration_compatible(&other.network_devices)
-            .map_err(|e| {
-                MigrationCompatibilityError::SpecMismatch(
-                    "network devices".to_string(),
-                    e,
-                )
-            })?;
-
-        self.network_backends
-            .is_migration_compatible(&other.network_backends)
-            .map_err(|e| {
-                MigrationCompatibilityError::SpecMismatch(
-                    "network backends".to_string(),
-                    e,
-                )
-            })?;
-
-        self.serial_ports
-            .is_migration_compatible(&other.serial_ports)
-            .map_err(|e| {
-                MigrationCompatibilityError::SpecMismatch(
-                    "serial ports".to_string(),
-                    e,
-                )
-            })?;
-
-        self.pci_pci_bridges
-            .is_migration_compatible(&other.pci_pci_bridges)
-            .map_err(|e| {
-                MigrationCompatibilityError::SpecMismatch(
-                    "PCI bridges".to_string(),
-                    e,
-                )
-            })?;
+        self.backends
+            .is_migration_compatible(&other.backends)
+            .map_err(|e| MigrationCompatibilityError::SpecMismatch(e.0, e.1))?;
 
         Ok(())
     }
